+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 - Fix typos on UR modules
 - Add `.prg` program init for ABB modules
 - Add `GetDeviceDriverModules(params)` as a way of retrieving module files
+- Quick and dirty fixing of the UR manager not being able to reconnect after disconnection
 
 
 ---